==> notebook/preprocess/0_preprocess_ice.ipynb <==
# extract CAS-RN and toxicity data from ICE
----------
## extract LD50 or DART test from ICE and extract CAS-RN and toxicity data 

import statistics
import pandas as pd

### LD50

df = pd.read_excel("../../data/raw/Acute_Oral_Toxicity.xlsx")

# use only chemical data
cas = dict()
unit = []
for i in range(len(df)):
    if df["Mixture"][i] != "Chemical":
        continue
    if df["CASRN"][i] not in cas.keys():
        cas[df["CASRN"][i]] = []
        cas[df["CASRN"][i]].append(df["Response"][i])
        unit.append(df["Response Unit"][i])
    else:
        cas[df["CASRN"][i]].append(df["Response"][i])
        unit.append(df["Response Unit"][i])

print(set(unit))

tsv = []
for i in cas.keys():
    col = []
    col.append(i)
    col.append(statistics.median(cas[i]))
    tsv.append(col)

pd.DataFrame(tsv).to_csv("../../data/processed/0701/cas_sev.tsv", sep="\t", header=None, index=False)
pd.DataFrame(tsv).to_csv("../../data/processed/0702/cas_sev.tsv", sep="\t", header=None, index=False)

### DART

df = pd.read_excel("../../data/raw/DART.xlsx")

df

cas = dict()
for i in range(len(df)):
    if str(df["Critical Effect"][i]) == "nan":
        continue
    if df["CASRN"][i] not in cas.keys():
        cas[df["CASRN"][i]] = []
        cas[df["CASRN"][i]].append(df["Critical Effect"][i])
    else:
        cas[df["CASRN"][i]].append(df["Critical Effect"][i])

tsv = []
for i in cas.keys():
    col = []
    if "Yes" in cas[i]:
       col.append(i)
       col.append(1)
    else:
        col.append(i)
        col.append(0)
    tsv.append(col)

pd.DataFrame(tsv).to_csv("../../data/processed/1001/cas_sev.tsv", sep="\t", header=None, index=False)
pd.DataFrame(tsv).to_csv("../../data/processed/1002/cas_sev.tsv", sep="\t", header=None, index=False)
