==> notebook/preprocess/1_preprocess_toxcast.ipynb <==
# extract CAS-RN and toxicity data from toxcast
----------
## extract tox21 ER or AR test from toxcast and extract CAS-RN and toxicity data 

import pandas as pd
from tqdm import tqdm

df = pd.read_csv("../../data/processed/toxcast/m5dat.tsv", sep="\t")

for rec in ["AR", "ERa", "ERb"]:
    for lig in ["Agonist", "Antagonist"]:
        set_assey = set()
        for i in range(len(df)):
            if "TOX21" in df["aenm"][i] and f"_{rec}_" in df["aenm"][i] and lig in df["aenm"][i]:
                set_assey.add(df["aenm"][i])

        before_tsv = []
        for i in tqdm(range(len(df))):
            if df["aenm"][i] in set_assey:
                col = []
                col.append(df["casn"][i])
                col.append(df["hitc"][i])
                before_tsv.append(col)
        
        cas_sev = dict()
        for i in range(len(before_tsv)):
            if before_tsv[i][0] == "nan":
                continue
            elif before_tsv[i][0] in cas_sev.keys():
                cas_sev[before_tsv[i][0]].add(before_tsv[i][1])
            elif before_tsv[i][0] not in cas_sev.keys():
                cas_sev[before_tsv[i][0]] = set()
                cas_sev[before_tsv[i][0]].add(before_tsv[i][1])

        tsv = []
        for cas in cas_sev.keys():
            col = []
            col.append(cas)
            if 1 in cas_sev[cas]:
                col.append(1)
            else:
                col.append(0)
            tsv.append(col)

        if lig == "Agonist":
            l = "ago"
        elif lig == "Antagonist":
            l = "anta"

        pd.DataFrame(tsv).to_csv(f"../../data/processed/tox21/tox21_{rec}_{l}.tsv", sep="\t", header=None, index=False)

        print(rec, lig, "DONE!")

## for validation test
--------
### Each jacvam validation test for endocrine disruption confirms the following different ligands.  
・jacvam_09_01 agonist    → ER (ERa + ERb) agonist  
・jacvam_09_01 antagonist → ER (ERa + ERb) antagonist  
・jacvam_09_02 agonist    → ERa agonist  
・jacvam_09_02 antagonist → ERa antagonist  
・jacvam_09_04 agonist    → AR agonist  
・jacvam_09_04 antagonist → AR antagonist  
・jacvam_09_05 agonist    → AR agonist  
・jacvam_09_06 antagonist → AR antagonist  
・jacvam_09_07            → ER (ERa + ERb) agonist and antagonist  

import os
import sys

current_dir = os.getcwd()
parent_parent_dir = os.path.dirname(os.path.dirname(current_dir))
src_dir = os.path.join(parent_parent_dir, 'src')
sys.path.append(src_dir)

from prep import prep_for_09

for test in ["01", "02", "04", "05"]:
    for lig in ["ago", "anta"]:
        prep_for_09(test, lig)
        print(test, lig, "DONE!")

prep_for_09("07")
print("07 DONE!")
