==> notebook/preprocess/2_preprocess_validation_test.ipynb <==
# make dataset for genetic algolism
-----
・make CASRN-severity dataset without validation test's compounds

import os
import sys

current_dir = os.getcwd()
parent_parent_dir = os.path.dirname(os.path.dirname(current_dir))
src_dir = os.path.join(parent_parent_dir, 'src')
sys.path.append(src_dir)

from prep import prep_for_ga

for test_num in ["0901", "0902", "0904", "0905"]:
    for lig in ["ago", "anta"]:
        print(test_num, lig)
        prep_for_ga(test_num, lig)

for test_num in ["0907", "0701", "0702", "1001", "1002"]:
    print(test_num)
    prep_for_ga(test_num)

# make dataset for validation test
----
・make CASRN-severity dataset with validation test's compounds  
・if validation test's compound's severity data is not in ICE or TOXCAST data, use jacvam's data (written in jacvam's text)

import os
import sys

current_dir = os.getcwd()
parent_parent_dir = os.path.dirname(os.path.dirname(current_dir))
src_dir = os.path.join(parent_parent_dir, 'src')
sys.path.append(src_dir)

from prep import prep_validation

for test_num in ["0901", "0902", "0904", "0905"]:
    for lig in ["ago", "anta"]:
        print(test_num, lig)
        prep_validation(test_num, lig)

for test_num in ["0907", "0701", "0702", "1001", "1002"]:
    print(test_num)
    prep_validation(test_num)

### get toxisity data about printed CASRN from jacvam text 
save in ../../data/row/validation_test_cas_not_in_whole

assign data in ../../data/row/validation_test_CAS_not_in_whole 

import pandas as pd
from pandas.errors import EmptyDataError

df = pd.read_csv("/workspace/HDD1/user/ohto_2/1-project/2-tox_pred/1-ga_mmo/github/multi-objective-optimization/data/processed/1002/cas_not_in_whole.tsv", sep="\t", header=None)

df.info()

for test_num in ["0901", "0902", "0904", "0905"]:
    for lig in ["ago", "anta"]:
        try:
            df_in_whole = pd.read_csv(f"../../data/processed/{test_num}_{lig}/validation.tsv", sep="\t", header=None)
            df_not_in_whole = pd.read_csv(f"../../data/processed/{test_num}_{lig}/cas_not_in_whole.tsv", sep="\t", header=None)

            df = pd.concat([df_in_whole, df_not_in_whole]).reset_index(drop=True)
            df.to_csv(f"../../data/processed/{test_num}_{lig}/validation.tsv", sep="\t", header=None, index=False)
        except EmptyDataError:
            continue

for test_num in ["0907", "0701", "0702", "1001", "1002"]:
    try:
        df_in_whole = pd.read_csv(f"../../data/processed/{test_num}/validation.tsv", sep="\t", header=None)
        df_not_in_whole = pd.read_csv(f"../../data/processed/{test_num}/cas_not_in_whole.tsv", sep="\t", header=None)

        df = pd.concat([df_in_whole, df_not_in_whole]).reset_index(drop=True)
        df.to_csv(f"../../data/processed/{test_num}/validation.tsv", sep="\t", header=None, index=False)
    except EmptyDataError:
        continue
